requirements.txt
pytz
pyyaml
feedgen
git+https://github.com/nixonjoshua98/manganelo.git#master
